# docs/reference/openapi.yaml
openapi: 3.1.0
info:
  title: Yantra4D API
  version: 1.0.0
  description: |
    REST API for the Yantra4D parametric 3D design platform.
    Provides OpenSCAD rendering, project management, AI-assisted design,
    Git operations, and analytics.
  license:
    name: AGPL-3.0
    url: https://opensource.org/license/agpl-v3/
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.yantra4d.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Janua-issued JWT token
  schemas:
    Error:
      type: object
      properties:
        status:
          type: string
          example: error
        error:
          type: string
      required: [status, error]
...
    RenderRequest:
      type: object
      properties:
        project:
          type: string
          description: Project slug
        mode:
          type: string
          description: Mode ID from manifest
        scad_file:
          type: string
          description: Deprecated — use mode instead
        parameters:
          type: object
          additionalProperties: true
        export_format:
          type: string
          enum: [stl, 3mf, off]
          default: stl
        ignore_cache:
          type: boolean
          description: Bypass L1 and L2 caches to force a fresh render
          default: false
      required: [mode]
    RenderResponse:
      type: object
      properties:
        status:
          type: string
          example: success
        parts:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              url:
                type: string
                format: uri
              size_bytes:
                type: integer
        log:
          type: string
    EstimateRequest:
      type: object
      properties:
        project:
          type: string
        mode:
          type: string
        scad_file:
          type: string
        parameters:
          type: object
    EstimateResponse:
      type: object
      properties:
        estimated_seconds:
          type: number
        num_parts:
          type: integer
        num_units:
          type: number
...
paths:
...
  /api/projects/{slug}/parts/{filename}:
    get:
      tags: [projects]
      summary: Serve static part file
      operationId: serveStaticPart
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
        - name: filename
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Binary STL file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "403":
          description: Path traversal blocked
        "404":
          description: File or project not found
...
  /api/estimate:
    post:
      tags: [render]
      summary: Estimate render time
      operationId: estimateRenderTime
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateRequest"
      responses:
        "200":
          description: Time estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EstimateResponse"
  /api/render:
    post:
      tags: [render]
      summary: Synchronous render
      operationId: renderStl
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderRequest"
      responses:
        "200":
          description: Render complete
          headers:
            X-RateLimit-Limit:
              schema:
                type: string
            X-RateLimit-Tier:
              schema:
                type: string
            X-Cache:
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RenderResponse"
        "400":
          description: Invalid mode or SCAD file
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/render-stream:
    post:
      tags: [render]
      summary: Stream render via SSE
      operationId: renderStlStream
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenderRequest"
      responses:
        "200":
          description: Server-Sent Events stream
          content:
            text/event-stream:
              schema:
                type: string
                description: |
                  SSE events: progress, part_done, complete.
                  Each `data:` line is a JSON object with an `event` field.
...